Extract tests into separate test classes
