+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/MiscTest.java
@@ -1462,34 +1462,6 @@ import org.slf4j.LoggerFactory;
      }
 
      @Test
-     public void testNodeSharingNotExists() throws Exception {
-         KnowledgeBase kbase = SerializationHelper.serializeObject( loadKnowledgeBase( "test_nodeSharingNotExists.drl" ) );
-         StatefulKnowledgeSession ksession = createKnowledgeSession( kbase );
-
-         final List list = new ArrayList();
-         ksession.setGlobal( "results",
-                             list );
-
-         ksession.fireAllRules();
-
-         assertEquals( 1,
-                       list.size() );
-
-         assertEquals( "rule1",
-                       list.get( 0 ) );
-
-         ksession.insert( new Cheese( "stilton",
-                                      10 ) );
-         ksession.fireAllRules();
-
-         assertEquals( 2,
-                       list.size() );
-
-         assertEquals( "rule2",
-                       list.get( 1 ) );
-     }
-
-     @Test
      public void testAutomaticBindings() throws Exception {
          final KnowledgeBase kbase = SerializationHelper.serializeObject( loadKnowledgeBase( "test_AutoBindings.drl" ) );
          final StatefulKnowledgeSession ksession = createKnowledgeSession( kbase );
@@ -1637,29 +1609,6 @@ import org.slf4j.LoggerFactory;
                            results.get( i ) );
          }
          results.clear();
-
-     }
-
-     @Test
-     public void testEqualitySupport() throws Exception {
-         final KnowledgeBase kbase = SerializationHelper.serializeObject( loadKnowledgeBase( "test_equalitySupport.drl" ) );
-         final StatefulKnowledgeSession ksession = createKnowledgeSession( kbase );
-
-         final List results = new ArrayList();
-         ksession.setGlobal( "results",
-                             results );
-
-         PersonWithEquals person = new PersonWithEquals( "bob",
-                                                         30 );
-
-         ksession.insert( person );
-
-         ksession.fireAllRules();
-
-         assertEquals( 1,
-                       results.size() );
-         assertEquals( "mark",
-                       results.get( 0 ) );
 
      }
 
@@ -2770,224 +2719,6 @@ import org.slf4j.LoggerFactory;
                        list.size() );
 
          ksession.dispose();
-     }
-
-     @Test
-     public void testLastMemoryEntryExistsBug() {
-         // JBRULES-2809
-         // This occurs when a blocker is the last in the node's memory, or if there is only one fact in the node
-         // And it gets no opportunity to rematch with itself
-
-         String str = "";
-         str += "package org.simple \n";
-         str += "import " + A.class.getCanonicalName() + "\n";
-         str += "global java.util.List list \n";
-         str += "rule x1 \n";
-         str += "when \n";
-         str += "    $s : String( this == 'x1' ) \n";
-         str += "    exists A( this != null ) \n";
-         str += "then \n";
-         str += "  list.add(\"fired x1\"); \n";
-         str += "end  \n";
-         str += "rule x2 \n";
-         str += "when \n";
-         str += "    $s : String( this == 'x2' ) \n";
-         str += "    exists A( field1 == $s, this != null ) \n"; // this ensures an index bucket
-         str += "then \n";
-         str += "  list.add(\"fired x2\"); \n";
-         str += "end  \n";
-
-         KnowledgeBase kbase = loadKnowledgeBaseFromString( str );
-
-         StatefulKnowledgeSession ksession = createKnowledgeSession( kbase );
-         List list = new ArrayList();
-         ksession.setGlobal( "list",
-                             list );
-
-         ksession.insert( "x1" );
-         ksession.insert( "x2" );
-         A a1 = new A( "x1",
-                       null );
-         A a2 = new A( "x2",
-                       null );
-
-         FactHandle fa1 = ksession.insert( a1 );
-         FactHandle fa2 = ksession.insert( a2 );
-
-         // make sure the 'exists' is obeyed when fact is cycled causing add/remove node memory
-         ksession.update( fa1,
-                          a1 );
-         ksession.update( fa2,
-                          a2 );
-         ksession.fireAllRules();
-
-         assertEquals( 2,
-                       list.size() );
-
-         ksession.dispose();
-     }
-
-     @Test
-     public void testNotIterativeModifyBug() {
-         // JBRULES-2809
-         // This bug occurs when a tuple is modified, the remove/add puts it onto the memory end
-         // However before this was done it would attempt to find the next tuple, starting from itself
-         // This meant it would just re-add itself as the blocker, but then be moved to end of the memory
-         // If this tuple was then removed or changed, the blocked was unable to check previous tuples.
-
-         String str = "";
-         str += "package org.simple \n";
-         str += "import " + A.class.getCanonicalName() + "\n";
-         str += "global java.util.List list \n";
-         str += "rule xxx \n";
-         str += "when \n";
-         str += "  $f1 : A() \n";
-         str += "    not A(this != $f1,  eval(field2 == $f1.getField2())) \n";
-         str += "    eval( !$f1.getField1().equals(\"1\") ) \n";
-         str += "then \n";
-         str += "  list.add($f1); \n";
-         str += "end  \n";
-
-         KnowledgeBase kbase = loadKnowledgeBaseFromString( str );
-
-         StatefulKnowledgeSession ksession = createKnowledgeSession( kbase );
-         List list = new ArrayList();
-         ksession.setGlobal( "list",
-                             list );
-
-         A a1 = new A( "2",
-                       "2" );
-         A a2 = new A( "1",
-                       "2" );
-         A a3 = new A( "1",
-                       "2" );
-
-         FactHandle fa1 = ksession.insert( a1 );
-         FactHandle fa2 = ksession.insert( a2 );
-         FactHandle fa3 = ksession.insert( a3 );
-         ksession.fireAllRules();
-
-         // a1 is blocked by a2
-         assertEquals( 0,
-                       list.size() );
-
-         // modify a2, so that a1 is now blocked by a3
-         a2.setField2( "1" ); // Do
-         ksession.update( fa2,
-                          a2 );
-         a2.setField2( "2" ); // Undo
-         ksession.update( fa2,
-                          a2 );
-
-         // modify a3 to cycle, so that it goes on the memory end, but in a previous bug still blocked a1
-         ksession.update( fa3,
-                          a3 );
-
-         a3.setField2( "1" ); // Do
-         ksession.update( fa3,
-                          a3 );
-         ksession.fireAllRules();
-         assertEquals( 0,
-                       list.size() ); // this should still now blocked by a2, but bug from previous update hanging onto blocked
-
-         ksession.dispose();
-     }
-
-     @Test
-     public void testExistsIterativeModifyBug() {
-         // JBRULES-2809
-         // This bug occurs when a tuple is modified, the remove/add puts it onto the memory end
-         // However before this was done it would attempt to find the next tuple, starting from itself
-         // This meant it would just re-add itself as the blocker, but then be moved to end of the memory
-         // If this tuple was then removed or changed, the blocked was unable to check previous tuples.
-
-         String str = "";
-         str += "package org.simple \n";
-         str += "import " + A.class.getCanonicalName() + "\n";
-         str += "global java.util.List list \n";
-         str += "rule xxx \n";
-         str += "when \n";
-         str += "  $f1 : A() \n";
-         str += "    exists A(this != $f1, eval(field2 == $f1.getField2())) \n";
-         str += "    eval( !$f1.getField1().equals(\"1\") ) \n";
-         str += "then \n";
-         str += "  list.add($f1); \n";
-         str += "end  \n";
-
-         KnowledgeBase kbase = loadKnowledgeBaseFromString( str );
-
-         StatefulKnowledgeSession ksession = createKnowledgeSession( kbase );
-         List list = new ArrayList();
-         ksession.setGlobal( "list",
-                             list );
-
-         A a1 = new A( "2",
-                       "2" );
-         A a2 = new A( "1",
-                       "2" );
-         A a3 = new A( "1",
-                       "2" );
-
-         FactHandle fa1 = ksession.insert( a1 );
-         FactHandle fa2 = ksession.insert( a2 );
-         FactHandle fa3 = ksession.insert( a3 );
-
-         // a2, a3 are blocked by a1
-         // modify a1, so that a1,a3 are now blocked by a2
-         a1.setField2( "1" ); // Do
-         ksession.update( fa1,
-                          a1 );
-         a1.setField2( "2" ); // Undo
-         ksession.update( fa1,
-                          a1 );
-
-         // modify a2, so that a1,a2 are now blocked by a3
-         a2.setField2( "1" ); // Do
-         ksession.update( fa2,
-                          a2 );
-         a2.setField2( "2" ); // Undo
-         ksession.update( fa2,
-                          a2 );
-
-         // modify a3 to cycle, so that it goes on the memory end, but in a previous bug still blocked a1
-         ksession.update( fa3,
-                          a3 );
-
-         a3.setField2( "1" ); // Do
-         ksession.update( fa3,
-                          a3 );
-         ksession.fireAllRules();
-         assertEquals( 1,
-                       list.size() ); // a2 should still be blocked by a1, but bug from previous update hanging onto blocked
-
-         ksession.dispose();
-     }
-
-     @Test
-     public void testNotEqualsOperator() {
-         // JBRULES-3003: restriction evaluation returns 'false' for "trueField != falseField"
-
-         String str = "package org.drools.compiler\n" +
-                      "rule NotEquals\n" +
-                      "when\n" +
-                      "    Primitives( booleanPrimitive != booleanWrapper )\n" +
-                      "then\n" +
-                      "end";
-
-         KnowledgeBase kbase = loadKnowledgeBaseFromString( str );
-         StatefulKnowledgeSession ksession = createKnowledgeSession( kbase );
-
-         Primitives p = new Primitives();
-         p.setBooleanPrimitive( true );
-         p.setBooleanWrapper( Boolean.FALSE );
-
-         ksession.insert( p );
-
-         int rules = ksession.fireAllRules();
-         ksession.dispose();
-
-         assertEquals( 1,
-                       rules );
      }
 
      @Test
@@ -4170,67 +3901,6 @@ import org.slf4j.LoggerFactory;
          verify( ael, times( 2 ) ).afterMatchFired(any(org.kie.api.event.rule.AfterMatchFiredEvent.class));
          // no cancellations should have happened
          verify( ael, never() ).matchCancelled( any( org.kie.api.event.rule.MatchCancelledEvent.class ) );
-     }
-
-     @Test
-     public void testDeterministicOTNOrdering() throws Exception {
-         // JBRULES-3632
-         String str =
-                 "package indexingproblem.remove.me.anditworks;\n" +
-                         "declare Criteria\n" +
-                         "   processed : boolean\n" +
-                         "end\n" +
-                         "\n" +
-                         "declare CheeseCriteria extends Criteria end\n" +
-                         "\n" +
-                         "rule setUp salience 10000 when\n" +
-                         "then\n" +
-                         "   insert(new CheeseCriteria());\n" +
-                         "end\n" +
-                         "\n" +
-                         "rule aaa when\n" +
-                         "   CheeseCriteria( )\n" +
-                         "then\n" +
-                         "end\n" +
-                         "\n" +
-                         "rule bbb when\n" +
-                         "   CheeseCriteria( )\n" +
-                         "then\n" +
-                         "end\n" +
-                         "\n" +
-                         "rule ccc when\n" +
-                         "   CheeseCriteria( )\n" +
-                         "then\n" +
-                         "end\n" +
-                         "\n" +
-                         "rule eeeFalse when\n" +
-                         "   Criteria( processed == false )\n" +
-                         "then\n" +
-                         "end\n" +
-                         "\n" +
-                         "declare Filter end\n" +
-                         "\n" +
-                         "rule fffTrue when\n" +
-                         "   Criteria( processed == true )\n" +
-                         "   Filter( )\n" +
-                         "then\n" +
-                         "end\n" +
-                         "\n" +
-                         "rule ruleThatFails when\n" +
-                         "   $criteria : Criteria( processed == false )\n" +
-                         "then\n" +
-                         "   modify($criteria) { setProcessed(true) }\n" +
-                         "end";
-
-         KnowledgeBase kbase = loadKnowledgeBaseFromString( str );
-         kbase = SerializationHelper.serializeObject( kbase );
-         StatefulKnowledgeSession ksession = kbase.newStatefulKnowledgeSession();
-
-         ksession.fireAllRules();
-
-         // check that OTNs ordering is not breaking serialization
-         ksession = SerializationHelper.getSerialisedStatefulKnowledgeSession( ksession, true );
-         ksession.fireAllRules();
      }
 
      @Test
