+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/operators/EvalRewriteTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.drools.compiler.integrationtests.operators.eval;
+package org.drools.compiler.integrationtests.operators;
 
 import java.util.ArrayList;
 import java.util.List;
