+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/operators/OrTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.drools.compiler.integrationtests.operators.or;
+package org.drools.compiler.integrationtests.operators;
 
 import java.io.IOException;
 import java.util.ArrayList;
