+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/session/ModifyTest.java
@@ -25,6 +25,7 @@ import org.drools.compiler.CommonTestMethodBase;
 import org.drools.compiler.IndexedNumber;
 import org.drools.compiler.OuterClass;
 import org.drools.compiler.Person;
+import org.drools.compiler.integrationtests.MiscTest;
 import org.drools.compiler.integrationtests.SerializationHelper;
 import org.junit.Assert;
 import org.junit.Test;
@@ -297,5 +298,60 @@ public class ModifyTest extends CommonTestMethodBase {
         setterList.add(CommandFactory.newSetter("likes", p.getLikes()));
 
         ksession.execute(CommandFactory.newModify(fh, setterList));
+    }
+
+    @Test
+    public void testNotIterativeModifyBug() {
+        // JBRULES-2809
+        // This bug occurs when a tuple is modified, the remove/add puts it onto the memory end
+        // However before this was done it would attempt to find the next tuple, starting from itself
+        // This meant it would just re-add itself as the blocker, but then be moved to end of the memory
+        // If this tuple was then removed or changed, the blocked was unable to check previous tuples.
+
+        String str = "";
+        str += "package org.simple \n";
+        str += "import " + MiscTest.A.class.getCanonicalName() + "\n";
+        str += "global java.util.List list \n";
+        str += "rule xxx \n";
+        str += "when \n";
+        str += "  $f1 : A() \n";
+        str += "    not A(this != $f1,  eval(field2 == $f1.getField2())) \n";
+        str += "    eval( !$f1.getField1().equals(\"1\") ) \n";
+        str += "then \n";
+        str += "  list.add($f1); \n";
+        str += "end  \n";
+
+        final KieBase kbase = loadKnowledgeBaseFromString(str);
+        final KieSession ksession = createKnowledgeSession(kbase);
+        final List list = new ArrayList();
+        ksession.setGlobal("list", list);
+
+        final MiscTest.A a1 = new MiscTest.A("2", "2");
+        final MiscTest.A a2 = new MiscTest.A("1", "2");
+        final MiscTest.A a3 = new MiscTest.A("1", "2");
+
+        final FactHandle fa1 = ksession.insert(a1);
+        final FactHandle fa2 = ksession.insert(a2);
+        final FactHandle fa3 = ksession.insert(a3);
+        ksession.fireAllRules();
+
+        // a1 is blocked by a2
+        assertEquals(0, list.size());
+
+        // modify a2, so that a1 is now blocked by a3
+        a2.setField2("1"); // Do
+        ksession.update(fa2, a2);
+        a2.setField2("2"); // Undo
+        ksession.update(fa2, a2);
+
+        // modify a3 to cycle, so that it goes on the memory end, but in a previous bug still blocked a1
+        ksession.update(fa3, a3);
+
+        a3.setField2("1"); // Do
+        ksession.update(fa3, a3);
+        ksession.fireAllRules();
+        assertEquals(0, list.size()); // this should still now blocked by a2, but bug from previous update hanging onto blocked
+
+        ksession.dispose();
     }
 }
