+++ drools-compiler/src/test/resources/org/drools/compiler/integrationtests/operators/eval_rule_test.drl
