+++ drools-compiler/src/test/resources/org/drools/compiler/integrationtests/operators/or_test.drl
