+++ drools-compiler/src/test/resources/org/drools/compiler/integrationtests/operators/test_EvalException.drl
