+++ drools-compiler/src/test/resources/org/drools/compiler/integrationtests/operators/test_contains_in_array.drl
